% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alberta.R
\name{theme_alberta}
\alias{theme_alberta}
\title{The University of Alberta ggplot2 theme}
\usage{
theme_alberta(colour_use, colour_palette)
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on the base plot.}

\item{colour_palette}{either 'alpha', 'beta', 'gamma' or 'delta' depending on which palette is selected.}
}
\value{
reformatted plot with the University of Alberta's visual identity theme.
}
\description{
theme_alberta() creates a theme using the University of Alberta visual identity
to be used with ggplot2. See the visual identity at https://www.ualberta.ca/toolkit/visual-identity/our-colours
}
\examples{

\dontrun{
ggplot(data=mtcars, aes(x = hp, y = mpg, colour = as.factor(cyl))) +
geom_point(size = 2.5) +
theme_alberta('colour', 'beta')
}

\dontrun{
ggplot(data = mtcars, aes(x = hp, fill = as.factor(cyl))) +
geom_density(alpha = 0.8) +
theme_alberta('fill', 'alpha')
}
}
