% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mcgill.R
\name{theme_mcgill}
\alias{theme_mcgill}
\title{McGill University Plotting Theme}
\usage{
: ggplot(data, aes(x, y, colour = class)) + theme_mcgill('colour')
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on how the colour will be displayed.}
}
\value{
reformatted plot with McGill University's visual identity theme.
}
\description{
: theme_mcgill() creates a theme using the McGill University's visual identity
to be used with ggplot2. See the visual identity at https://www.mcgill.ca/visual-identity/visual-identity-guide
}
\examples{

\dontrun{ggplot(iris, aes(Sepal.Length, Sepal.Width, fill = as.factor(Species))) + geom_boxplot() + theme_mcgill('fill')}

\dontrun{ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = as.factor(Species))) + geom_point() + theme_mcgill('colour')}
}
