% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_toronto.R
\name{theme_toronto}
\alias{theme_toronto}
\title{The University of Toronto ggplot2 theme}
\usage{
theme_toronto(colour_use, colour_palette)
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on the base plot.}

\item{colour_palette}{either 'vibrant', 'cool' or 'awards' depending on the use of the visualization.}
}
\value{
reformatted plot with the University of Toronto's visual identity theme.
}
\description{
: theme_toronto() creates a theme using the University of Toronto visual identity
to be used with ggplot2. See the visual identity at https://tinyurl.com/t3jjr49
}
\examples{

\dontrun{
ggplot(data = mtcars, aes(x = hp, y = mpg, colour = as.factor(cyl))) +
geom_point(size = 2.5) +
theme_toronto('colour', 'vibrant')
}

\dontrun{
ggplot(data = mtcars, aes(x = hp, fill = as.factor(cyl))) +
geom_density(alpha = 0.8) +
theme_toronto('fill', 'awards')
}
}
