% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ubc.R
\name{theme_ubc}
\alias{theme_ubc}
\title{The University of British Columbia ggplot2 theme}
\usage{
theme_ubc(colour_use)
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on the base plot.}
}
\value{
reformatted plot with the University of British Columbia's visual identity theme.
}
\description{
theme_ubc() creates a theme using the University of British Columbia's visual identity
to be used with ggplot2. See the visual identity at https://brand.ubc.ca/guidelines/downloads/ubc-colour-and-fonts/
}
\examples{

\dontrun{
ggplot(data = mtcars, aes(x = hp, y = mpg, colour = as.factor(cyl))) +
geom_point(size = 2.5) +
theme_ubc('colour')
}

\dontrun{
ggplot(data = mtcars, aes(x = hp, fill = as.factor(cyl))) +
geom_density(alpha = 0.8) +
theme_ubc('fill')
}
}
