% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_toronto.R
\name{theme_toronto}
\alias{theme_toronto}
\title{University of Toronto Plotting Theme}
\usage{
: ggplot(data, aes(x, y, colour = class)) + theme_toronto('colour', 'vibrant')
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on how the colour will be displayed.}

\item{colour_palette}{either 'vibrant', 'cool' or 'awards' depending on the use of the visualization.}
}
\value{
reformatted plot with University of Toronto's visual identity theme.
}
\description{
: theme_toronto() creates a theme using the University of Toronto visual identity
to be used with ggplot2. See the visual identity at https://tinyurl.com/t3jjr49
}
\examples{

\dontrun{ggplot(iris, aes(Sepal.Length, Sepal.Width, fill = as.factor(Species))) + geom_boxplot() + theme_toronto('fill', 'vibrant')}

\dontrun{ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = as.factor(Species))) + geom_point() + theme_toronto('colour', 'vibrant')}
}
