% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mcgill.R, R/theme_ubc.R
\name{theme_ubc}
\alias{theme_ubc}
\title{McGill University Plotting Theme}
\usage{
: ggplot(data, aes(x, y, colour = class)) + theme_ubc('colour')

: ggplot(data, aes(x, y, colour = class)) + theme_ubc('colour')
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on how the colour will be displayed.}
}
\value{
reformatted plot with UBC's visual identity theme.

reformatted plot with UBC's visual identity theme.
}
\description{
: theme_ubc() creates a theme using the University of British Columbia's visual identity
to be used with ggplot2. See the isual identity at https://brand.ubc.ca/guidelines/downloads/ubc-colour-and-fonts/

: theme_ubc() creates a theme using the University of British Columbia's visual identity
to be used with ggplot2. See the visual identity at https://brand.ubc.ca/guidelines/downloads/ubc-colour-and-fonts/
}
\examples{

\dontrun{ggplot(students_admissions, aes(x = high_gpa, y = department, fill = as.factor(department)))
     + geom_bar(stat = 'identity')
     + theme_ubc(colour_use = 'fill')}

\dontrun{ggplot(students_admissions, aes(x = high_gpa, y = department, colour = as.factor(department)))
     + geom_bar(stat = 'identity')
     + theme_ubc(colour_use = 'colour')}

\dontrun{ggplot(students_admissions, aes(x = high_gpa, y = department, fill = as.factor(department)))
     + geom_bar(stat = 'identity')
     + theme_ubc(colour_use = 'fill')}

\dontrun{ggplot(students_admissions, aes(x = high_gpa, y = department, colour = as.factor(department)))
     + geom_bar(stat = 'identity')
     + theme_ubc(colour_use = 'colour')}
}
