% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ubc.R
\name{theme_ubc}
\alias{theme_ubc}
\title{University of British Columbia Plotting Theme}
\usage{
: ggplot(data, aes(x, y, colour = class)) + theme_ubc('colour')
}
\arguments{
\item{colour_use}{either 'fill' or 'colour' depending on how the colour will be displayed.}
}
\value{
reformatted plot with UBC's visual identity theme.
}
\description{
: theme_ubc() creates a theme using the University of British Columbia's visual identity
to be used with ggplot2. See the visual identity at https://brand.ubc.ca/guidelines/downloads/ubc-colour-and-fonts/
}
\examples{

\dontrun{ggplot(iris, aes(Sepal.Length, Sepal.Width, fill = as.factor(Species))) + geom_boxplot() + theme_ubc('fill')}

\dontrun{ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = as.factor(Species))) + geom_point() + theme_ubc('colour')}
}
